Force push changes that improve translations
