+++ src/languages/translations.ts
@@ -1,7 +1,7 @@
 import en from './en';
 import es from './es';
 import esES from './es-ES';
-import type {TranslationBase, TranslationFlatObject} from './types';
+import type {FlatTranslationsObject, TranslationDeepObject} from './types';
 
 /**
  * Converts an object to it's flattened version.
@@ -12,10 +12,10 @@ import type {TranslationBase, TranslationFlatObject} from './types';
  */
 // Necessary to export so that it is accessible to the unit tests
 // eslint-disable-next-line rulesdir/no-inline-named-export
-export function flattenObject(obj: TranslationBase): TranslationFlatObject {
+export function flattenObject(obj: TranslationDeepObject): FlatTranslationsObject {
     const result: Record<string, unknown> = {};
 
-    const recursive = (data: TranslationBase, key: string): void => {
+    const recursive = (data: TranslationDeepObject, key: string): void => {
         // If the data is a function or not a object (eg. a string or array),
         // it's the final value for the key being built and there is no need
         // for more recursion
@@ -27,7 +27,7 @@ export function flattenObject(obj: TranslationBase): TranslationFlatObject {
             // Recursive call to the keys and connect to the respective data
             Object.keys(data).forEach((k) => {
                 isEmpty = false;
-                recursive(data[k] as TranslationBase, key ? `${key}.${k}` : k);
+                recursive(data[k] as TranslationDeepObject, key ? `${key}.${k}` : k);
             });
 
             // Check for when the object is empty but a key exists, so that
@@ -39,7 +39,7 @@ export function flattenObject(obj: TranslationBase): TranslationFlatObject {
     };
 
     recursive(obj, '');
-    return result as TranslationFlatObject;
+    return result as FlatTranslationsObject;
 }
 
 export default {
