+++ tests/unit/TranslateTest.ts
@@ -2,7 +2,7 @@
 import CONFIG from '@src/CONFIG';
 import CONST from '@src/CONST';
 import * as translations from '@src/languages/translations';
-import type {TranslationFlatObject, TranslationPaths} from '@src/languages/types';
+import type {FlatTranslationsObject, TranslationPaths} from '@src/languages/types';
 import * as Localize from '@src/libs/Localize';
 import asMutable from '@src/types/utils/asMutable';
 import arrayDifference from '@src/utils/arrayDifference';
@@ -59,10 +59,10 @@ describe('translate', () => {
 });
 
 describe('Translation Keys', () => {
-    function traverseKeyPath(source: TranslationFlatObject, path?: string, keyPaths?: string[]): string[] {
+    function traverseKeyPath(source: FlatTranslationsObject, path?: string, keyPaths?: string[]): string[] {
         const pathArray = keyPaths ?? [];
         const keyPath = path ? `${path}.` : '';
-        (Object.keys(source) as Array<keyof TranslationFlatObject>).forEach((key) => {
+        (Object.keys(source) as TranslationPaths[]).forEach((key) => {
             if (typeof source[key] === 'object' && typeof source[key] !== 'function') {
                 // @ts-expect-error - We are modifying the translations object for testing purposes
                 traverseKeyPath(source[key], keyPath + key, pathArray);
@@ -126,7 +126,6 @@ describe('flattenObject', () => {
                         none: 'No description',
                     },
                     content: func,
-                    messages: ['Hello', 'Hi', 'Sup!'],
                 },
             },
         };
@@ -141,7 +140,6 @@ describe('flattenObject', () => {
             'complex.report.title.task': 'Task',
             'complex.report.description.none': 'No description',
             'complex.report.content': func,
-            'complex.report.messages': ['Hello', 'Hi', 'Sup!'],
         });
     });
 });
